+++ src/otx/algo/detection/layers/__init__.py
@@ -3,6 +3,8 @@
 """Custom layer implementations for detection task."""
 
 from .channel_attention_layer import ChannelAttention
+from .common_layer import AConv, ADown, Concat
 from .csp_layer import CSPLayer, CSPRepLayer
+from .elan_layer import ELAN, SPPELAN, RepNCSPELAN
 
-__all__ = ["CSPLayer", "ChannelAttention", "CSPRepLayer"]
+__all__ = ["CSPLayer", "ChannelAttention", "CSPRepLayer", "Concat", "ELAN", "RepNCSPELAN", "SPPELAN", "AConv", "ADown"]
